remove type check for primitive array values
Currently, when a request body is specified as an array of primitive values, the AbstractBodyContentGenerator will add an array type check for the first item of an array.
This results in a client that never sends a request body when you pass an array of primitive values. This fix changes it to a simpler `is_array` check.